stability: implement node liveness; first step towards new range leases
This change adds a node liveness table as a global system table.
Nodes periodically write updates to their liveness record by doing
a conditional put to the liveness table. The leader of the range
containing the node liveness table gossips the latest information
to the rest of the system.

Each node has a `NodeLiveness` object which can be used to query
the status of any other node to find out if it's live or non-live
according to the liveness threshold duration compared to the last
time it successfully heartbeat its liveness record.

The as-yet-unused `IncrementEpoch` mechanism is also added in this
PR, for eventual use with the planned epoch-based range leader leases.

Updated the range leader lease RFC to reflect current thinking.

+++ server/updates.go
@@ -243,7 +243,7 @@ func (s *Server) maybeReportUsage(running time.Duration) time.Duration {
 		return updateCheckFrequency
 	}
 	return s.maybeRunPeriodicCheck("metrics reporting",
-		keys.NodeLastUsageReportKey(int32(s.node.Descriptor.NodeID)), s.reportUsage)
+		keys.NodeLastUsageReportKey(s.node.Descriptor.NodeID), s.reportUsage)
 }
 
 func (s *Server) getReportingInfo() reportingInfo {
